% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCropAreaFPU.R
\name{calcCropAreaFPU}
\alias{calcCropAreaFPU}
\title{FPU Level Crop area}
\usage{
calcCropAreaFPU(gdx, yrs = NULL, crop = NULL)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}

\item{yrs}{Years to subset. Defaults to NULL for all years.}

\item{crop}{Crops to subset. Defaults to NULL for all crops.}
}
\value{
FPU level area_fpu as sf object
}
\description{
FPU Level Crop area
}
\examples{
\dontrun{x <- calcCropAreaFPU(gdx)}
}
\author{
Abhijeet Mishra
}
