% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProductionFPU.R
\name{calcProductionFPU}
\alias{calcProductionFPU}
\title{FPU Level Production}
\usage{
calcProductionFPU(gdx, yrs = NULL, crop = NULL)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
FPU level production as sf object
}
\description{
FPU Level Production
}
\examples{
\dontrun{x <- calcProductionFPU(gdx)}
}
\author{
Abhijeet Mishra
}
