% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYieldFPU.R
\name{calcYieldFPU}
\alias{calcYieldFPU}
\title{FPU Level Production}
\usage{
calcYieldFPU(gdx, yrs = NULL, crop = NULL)
}
\arguments{
\item{gdx}{final GDX from an IMPACT run}
}
\value{
FPU level production as sf object
}
\description{
FPU Level Production
}
\examples{
\dontrun{x <- calcYieldFPU(gdx)}
}
\author{
Abhijeet Mishra
}
