% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatial.R
\name{plotSpatial}
\alias{plotSpatial}
\title{Plotter for IMPACT results at FPU level}
\usage{
plotSpatial(
  df,
  wrap = NULL,
  ncol = NULL,
  scale = 1,
  plot_title = NULL,
  legend_title = NULL
)
}
\arguments{
\item{df}{sf dataframe object. Usually an output from from readFPU()}

\item{wrap}{If the plot needs to be wrapped e.g., in case plotting multiple
indicators. For example, use "fctr" if you want to plot irrigated and rainfed
areas in case area is being plotted}

\item{ncol}{Number of columns to be plotted. Defaults to NULL.}

\item{scale}{Numeric value to scale output. Use 1e3 for thousand to million
conversion for example. Defaults to 1.}

\item{plot_title}{Title of the plot. Defaults to NULL.}

\item{legend_title}{Title of the legend, Defaults to NULL.}
}
\value{
ggplot object with FPU level IMPACT results
}
\description{
Plotter for IMPACT results at FPU level
}
\examples{
\dontrun{
plotSpatial(df)
}
}
