% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSHP.R
\name{readSHP}
\alias{readSHP}
\title{Shapefile reader for IMPACT FPUs}
\usage{
readSHP()
}
\value{
Terra object with shapefile constaining FPUs for IMPACT
}
\description{
Shapefile reader for IMPACT FPUs
}
\examples{
\dontrun{
readSHP()
}
}
